Move dialogs initialization to workspace
